# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_auth.DeleteAccountInput:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  internal_auth.ForgotPasswordInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_auth.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.RegisterInput:
    properties:
      email:
        type: string
      nickname:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.ResetPasswordInput:
    properties:
      new_password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  internal_cart.AddItemReq:
    properties:
      quantity:
        minimum: 1
        type: integer
      sku_id:
        type: integer
    required:
    - quantity
    - sku_id
    type: object
  internal_cart.UpdateQtyReq:
    properties:
      quantity:
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
  internal_catalog.Product:
    properties:
      allergens:
        type: string
      buy_count:
        type: integer
      category_id:
        type: integer
      click_count:
        type: integer
      created_at:
        type: string
      description:
        type: string
      description_en:
        type: string
      id:
        type: integer
      image_url:
        type: string
      ingredients:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      name_en:
        type: string
      scenes:
        type: string
      shelf_days:
        type: integer
      skus:
        items:
          $ref: '#/definitions/internal_catalog.SKU'
        type: array
      tags:
        type: string
      taste:
        type: string
      updated_at:
        type: string
      weight_gram:
        type: integer
    type: object
  internal_catalog.SKU:
    properties:
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      price:
        type: integer
      product_id:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: AI-powered snack e-commerce platform with hybrid search, conversational
    agent, and Stripe payments.
  title: Kotoha API
  version: "1.0"
paths:
...
  /admin/products:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_catalog.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_catalog.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a product (admin)
      tags:
      - admin
  /admin/products/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product (admin)
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_catalog.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_catalog.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a product (admin)
      tags:
      - admin
...
  /auth/account:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Current password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.DeleteAccountInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete account (soft delete)
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ForgotPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Request password reset email
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ResetPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Reset password with token
      tags:
      - auth
  /cart:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's cart
      tags:
      - cart
  /cart/items:
    post:
      consumes:
      - application/json
      parameters:
      - description: SKU ID and quantity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_cart.AddItemReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - cart
  /cart/items/{skuID}:
    delete:
      parameters:
      - description: SKU ID
        in: path
        name: skuID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - cart
    put:
      consumes:
      - application/json
      parameters:
      - description: SKU ID
        in: path
        name: skuID
        required: true
        type: integer
      - description: New quantity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_cart.UpdateQtyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - cart
...
  /catalog/products:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List products with pagination
      tags:
      - catalog
  /catalog/products/{id}:
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_catalog.Product'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get product by ID
      tags:
      - catalog
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"